Count hits... string searches appear to be broken for now, unsure why, maybe something goes wrong when allocating ram for the strings..
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@6373 a1c6a512-1295-4272-9138-f99709370657

+++ apps/plugins/searchengine/token.h
@@ -51,7 +51,7 @@
 struct token {
       unsigned char kind;
       char spelling[256];
-      int intvalue;
+      long intvalue;
 };
 
 char *getstring(struct token *token);
